refactor: Move DragDropManager from Window to InputManager
