+++ src/Uno.UI/UI/Xaml/Internal/ContentRoot.uno.cs
@@ -2,6 +2,7 @@
 using System;
 using System.Diagnostics;
 using System.Linq;
+using Windows.UI.Core;
 
 namespace Uno.UI.Xaml.Core;
 
@@ -23,4 +24,6 @@ internal partial class ContentRoot
 			Debug.Assert(_host == host);
 		}
 	}
+
+	internal CoreDispatcher Dispatcher => CoreDispatcher.Main; //TODO:MZ: set dispatcher per content root (from SetHost?)
 }
