+++ src/Uno.UI/UI/Xaml/Internal/InputManager.cs
@@ -22,6 +22,8 @@ internal partial class InputManager : IInputInjectorTarget
 		ConstructKeyboardManager();
 
 		ConstructPointerManager();
+
+		InitDragAndDrop();
 	}
 
 	partial void ConstructKeyboardManager();
